Extracting ContrastCheckerMessage component
