+++ packages/block-editor/src/components/contrast-checker/contrast-checker-message.js
@@ -1,0 +1,58 @@
+/**
+ * WordPress dependencies
+ */
+import { speak } from '@wordpress/a11y';
+import { __ } from '@wordpress/i18n';
+import { Notice } from '@wordpress/components';
+import { useEffect } from '@wordpress/element';
+
+export default function ContrastCheckerMessage( {
+	colordBackgroundColor,
+	colordTextColor,
+	colordLinkColor,
+	backgroundColor,
+	textColor,
+	linkColor,
+	shouldShowTransparencyWarning,
+} ) {
+	let msg = '';
+	if ( shouldShowTransparencyWarning ) {
+		msg = __( 'Transparent text may be hard for people to read.' );
+	} else {
+		const backgroundColorBrightness = colordBackgroundColor.brightness();
+		msg =
+			( colordTextColor &&
+				backgroundColorBrightness < colordTextColor.brightness() ) ||
+			( colordLinkColor &&
+				backgroundColorBrightness < colordLinkColor.brightness() )
+				? __(
+						'This color combination may be hard for people to read. Try using a darker background color and/or a brighter text color.'
+				  )
+				: __(
+						'This color combination may be hard for people to read. Try using a brighter background color and/or a darker text color.'
+				  );
+	}
+
+	// Note: The `Notice` component can speak messages via its `spokenMessage`
+	// prop, but the contrast checker requires granular control over when the
+	// announcements are made. Notably, the message will be re-announced if a
+	// new color combination is selected and the contrast is still insufficient.
+	useEffect( () => {
+		const speakMsg = shouldShowTransparencyWarning
+			? __( 'Transparent text may be hard for people to read.' )
+			: __( 'This color combination may be hard for people to read.' );
+		speak( speakMsg );
+	}, [ backgroundColor, textColor, linkColor ] );
+
+	return (
+		<div className="block-editor-contrast-checker">
+			<Notice
+				spokenMessage={ null }
+				status="warning"
+				isDismissible={ false }
+			>
+				{ msg }
+			</Notice>
+		</div>
+	);
+}
